test(ingest/kinesis): use standard docker_compose_runner + in-process ingest
Refactor the Kinesis integration test to the repo's standard harness instead of
the bespoke subprocess strategy:
- Use the docker_compose_runner fixture + wait_for_port (as delta_lake/neo4j/
  clickhouse do) instead of shelling out to `docker compose up/down` with a hand-
  rolled health poll.
- Seed test data in-process via boto3 fixtures; delete the standalone
  setup_test_data.py that was invoked via subprocess.
- Run ingestion in-process via run_datahub_cmd instead of subprocess to the
  `datahub` CLI — removes the DATAHUB_BIN/"stale global datahub" workaround.
- File sink writes a relative path (run_datahub_cmd isolates to tmp_path); add an
  explicit container_name so wait_for_port can exec the LocalStack container.

Golden output is byte-identical; integration test passes against LocalStack 3.0.
