+++ metadata-ingestion/tests/integration/kinesis/docker-compose.yml
@@ -2,6 +2,8 @@ version: "3.8"
 services:
   localstack:
     image: localstack/localstack:3.0
+    # Explicit container_name so wait_for_port can `docker exec` it by name.
+    container_name: localstack
     ports:
       - "4566:4566"
     environment:
